refactored everything to do with a job page into it's own class
